+++ packages/@expo/metro-config/src/ExpoMetroConfig.ts
@@ -255,6 +255,7 @@ export function getDefaultConfig(
 
   const serverRoot = getMetroServerRoot(projectRoot);
 
+  const routerPackageRoot = resolveFrom.silent(projectRoot, 'expo-router');
   // Merge in the default config from Metro here, even though loadConfig uses it as defaults.
   // This is a convenience for getDefaultConfig use in metro.config.js, e.g. to modify assetExts.
   const metroConfig: Partial<MetroConfig> = mergeConfig(metroDefaultValues, {
@@ -351,6 +352,7 @@ export function getDefaultConfig(
       // Custom: These are passed to `getCacheKey` and ensure invalidation when the version changes.
       unstable_renameRequire: false,
       // @ts-expect-error: not on type.
+      _expoRouterPath: routerPackageRoot ? path.relative(serverRoot, routerPackageRoot) : undefined,
       postcssHash: getPostcssConfigHash(projectRoot),
       browserslistHash: pkg.browserslist
         ? stableHash(JSON.stringify(pkg.browserslist)).toString('hex')
@@ -364,7 +366,7 @@ export function getDefaultConfig(
       unstable_allowRequireContext: true,
       allowOptionalDependencies: true,
       babelTransformerPath: require.resolve('./babel-transformer'),
-      // TODO: The absolute path invalidates caching across devices.
+      // TODO: The absolute path invalidates caching across devices. To account for this, we remove the `asyncRequireModulePath` from the cache key but that means any changes to the file will not invalidate the cache.
       asyncRequireModulePath: require.resolve('./async-require'),
       assetRegistryPath: '@react-native/assets-registry/registry',
       // hermesParser: true,