+++ packages/expo-dev-menu/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+## 7.0.3 — 2025-08-18
+
+### 💡 Others
+
 - [ios] Wrap system color references for dev client. ([#38912](https://github.com/expo/expo/pull/38912) by [@douglowder](https://github.com/douglowder))
 
 ## 7.0.2 — 2025-08-16