+++ packages/expo-router/CHANGELOG.md
@@ -10,6 +10,19 @@
 
 ### 💡 Others
 
+## 6.0.0-beta.3 — 2025-08-18
+
+### 🐛 Bug fixes
+
+- remove error when link with preview is used for external link ([#38939](https://github.com/expo/expo/pull/38939) by [@Ubax](https://github.com/Ubax))
+- [Internal] Remove internal `SafeAreaViewSlot` to remove `SafeAreaView` import warning ([#38899](https://github.com/expo/expo/pull/38899) by [@kitten](https://github.com/kitten))
+- improve animation for stack push from different tab ([#38927](https://github.com/expo/expo/pull/38927) by [@Ubax](https://github.com/Ubax))
+
+### 💡 Others
+
+- Throw an error when tabs are nested ([#38651](https://github.com/expo/expo/pull/38651) by [@Ubax](https://github.com/Ubax))
+- refactor internal params logic ([#38929](https://github.com/expo/expo/pull/38929) by [@Ubax](https://github.com/Ubax))
+
 ## 6.0.0-beta.2 — 2025-08-16
 
 ### 🛠 Breaking changes